requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.30
python-dotenv
